ClnnModes: isvalidlnn_symmetric should only be symmetric if S=true
When S=true, we often (but not always) want to get the number when n2<n1 from
the number that is stored with n1<n2. However, when S=false, then the two
numbers need to be assumed to be different. Therefore, it is very hard to come
up with a situation where we would actually want to use isvalid_symmetric, and
it should fall back to isvalidlnn without the symmetry assumption.

Sure, it is a little awkward to have it this way, but it simplifies code, so
let's do it.